chore(browseros): move server_binaries into products, r2 into lib
