+++ packages/browseros/bos_build/cli/ota.py
@@ -18,7 +18,7 @@ from ..steps.ota.common import (
     get_appcast_path,
     SERVER_PLATFORMS,
 )
-from ..steps.storage import get_r2_client, upload_file_to_r2
+from ..lib.r2 import get_r2_client, upload_file_to_r2
 
 app = typer.Typer(
     help="OTA (Over-The-Air) update automation",
