+++ packages/browseros/bos_build/products/browseros/product.py
@@ -4,7 +4,7 @@
 from pathlib import Path
 
 from ...core.products import ProductDescriptor
-from ...core.server_binaries import ServerBundle, SignSpec
+from ..server_binaries import ServerBundle, SignSpec
 
 BROWSEROS_PRODUCT = ProductDescriptor.define(
     id="browseros",
